+++ src/print_dats.F90
@@ -8,7 +8,7 @@ SUBROUTINE print_dats()
     use com_senga
     use com_ops_senga
 
-    integer :: ispec
+    integer(4) :: ispec
     CHARACTER (LEN=60) :: fname
     CHARACTER (LEN=3) :: pnxres
     PARAMETER(pnxres = '.h5')
